bsdd updates
